Extract document methods into `Meilisearch::Index::Documents` module
